Move to cfgpath package
